moving from xapi to xo-server, optimization, cache handling
