+++ packages/xo-web/src/common/xo/index.js
@@ -754,7 +754,7 @@ export const subscribeIpmiSensors = host => {
   return subscribeHostsIpmiSensors[hostId]
 }
 
-export const check2crsiHostBiosVersion = host => _call('host.getBiosInfo', { id: resolveId(host) })
+export const getBiosInfo = host => _call('host.getBiosInfo', { id: resolveId(host) })
 
 const subscribeVmSecurebootReadiness = {}
 export const subscribeSecurebootReadiness = id => {
